backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
branca==0.8.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.116.1
filelock==3.19.1
folium==0.20.0
fonttools==4.59.2
fpdf==1.7.2
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.6
networkx==3.5
nltk==3.9.1
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2025.9.1
requests==2.32.5
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
snscrape==0.7.0.20230622
soupsieve==2.8
starlette==0.47.3
textblob==0.19.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wordcloud==1.9.4
xyzservices==2025.4.0

fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
